Encryption power loss...
This passage seems fatal: "Adam encrypts the message using AES 256bit GCM using the derived shared secret as the key and sends it to cherie, the to and from key version used to generate the message are included as part of the message." 

I know that we have an SSL layer over top of this, but, if that fails, can a key agreement be shared with a third party and maintain security? 

Is the key agreement digested in a deterministic way before it is "leaves the nest" and is transmitted? 

I note that surespot's use of GCM might also mitigate it, and as an AEAD mode it refuses chosen-ciphertext attacks. 

 - v6